+++ app/components/ContextMenuItem/index.jsx
@@ -20,7 +20,15 @@ const MOUSE_LEAVE_DELAY_MS = 150
 
 export const ContextMenuItem = ({ text, children = [], onClick = () => {} }) => {
   const elRef = React.useRef()
+
+  /*
+  Delayed hover is used for preventing unmounting of
+  child menus when the cursor leaves the item,
+  regular hover is used for tinting the item
+  */
+  const [delayedHover, setDelayedHover] = React.useState(false)
   const [hover, setHover] = React.useState(false)
+
   const childArr = Array.isArray(children) ? children : [children]
 
   const timeoutRef = React.useRef()
@@ -29,12 +37,14 @@ export const ContextMenuItem = ({ text, children = [], onClick = () => {} }) => 
     if (timeoutRef.current) {
       clearTimeout(timeoutRef.current)
     }
+    setDelayedHover(true)
     setHover(true)
   }
 
   function handleMouseLeave () {
+    setHover(false)
     timeoutRef.current = setTimeout(() => {
-      setHover(false)
+      setDelayedHover(false)
     }, MOUSE_LEAVE_DELAY_MS)
   }
 
@@ -45,6 +55,7 @@ export const ContextMenuItem = ({ text, children = [], onClick = () => {} }) => 
   }
 
   function handleFocus (e) {
+    setDelayedHover(true)
     setHover(true)
   }
 
@@ -53,7 +64,7 @@ export const ContextMenuItem = ({ text, children = [], onClick = () => {} }) => 
   return (
     <div
       ref={elRef}
-      className='ContextMenuItem'
+      className={`ContextMenuItem ${hover ? 'is-hovered' : ''}`}
       onMouseEnter={handleMouseEnter}
       onMouseLeave={handleMouseLeave}
       onClick={() => onClick()}
@@ -69,7 +80,7 @@ export const ContextMenuItem = ({ text, children = [], onClick = () => {} }) => 
         <Icon name='arrowRight' color='black' />
       }
       {
-        hover && childArr.length > 0
+        delayedHover && childArr.length > 0
           ? (
             <ContextMenu x={bounds?.x + CTX_MENU_OFFSET_X_PX} y={bounds?.y + bounds?.height / 2}>
               {children}